requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py38h0a891b7_1
aiosignal=1.2.0=pyhd8ed1ab_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.5.18.1=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
catalogue=2.0.6=py38h578d9bd_2
certifi=2022.5.18.1=py38h578d9bd_0
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py38h578d9bd_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
cymem=2.0.6=py38hfa26641_3
cython-blis=0.7.7=py38h71d37f0_1
dataclasses=0.8=pyhc8e2a94_3
de-core-news-sm=3.3.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
en-core-web-sm=3.3.0=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=py38h0a891b7_1
fsspec=2022.3.0=pyhd8ed1ab_0
future=0.18.2=py38h578d9bd_5
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.6=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py38hce63b2e_0
icu=67.1=he1b5a44_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38h578d9bd_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.13.0=pypi_0
ipython=8.3.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jupyter-client=7.3.1=pypi_0
jupyter-core=4.10.0=pypi_0
kiwisolver=1.4.2=py38h43d8883_1
lame=3.100=h7b6447c_0
langcodes=3.3.0=pyhd8ed1ab_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_0
libgomp=11.2.0=h1234567_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_1
libstdcxx-ng=11.2.0=h1234567_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_1
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h5d868c9_1
matplotlib-inline=0.1.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h0a891b7_1
murmurhash=1.0.7=py38hfa26641_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pypi_0
nettle=3.7.3=hbbd107a_1
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38h47df419_1
parso=0.8.3=pypi_0
pathy=0.6.1=pyhd8ed1ab_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
preshed=3.0.6=py38hfa26641_2
prompt-toolkit=3.0.29=pypi_0
protobuf=3.15.8=py38h709712a_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py38h497a2fe_2
pydeprecate=0.3.2=pyhd8ed1ab_0
pygments=2.12.0=pypi_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.6.3=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=23.0.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
setuptools=59.5.0=py38h578d9bd_0
shellingham=1.4.0=pyh44b312d_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=pyhd8ed1ab_0
spacy=3.3.0=py38h514daf8_0
spacy-legacy=3.0.9=pyhd8ed1ab_0
spacy-loggers=1.0.2=pyhd8ed1ab_0
sqlite=3.38.3=hc218d9a_0
srsly=2.4.3=py38hfa26641_1
stack-data=0.2.0=pypi_0
tensorboard=2.9.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38h2b5fc30_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
thinc=8.0.15=py38h514daf8_0
tk=8.6.11=h1ccaba5_1
torchdata=0.3.0=py38
torchmetrics=0.8.2=pyhd8ed1ab_0
torchtext=0.12.0=py38
torchvision=0.12.0=py38_cu113
tornado=6.1=py38h0a891b7_3
tqdm=4.64.0=py38h06a4308_0
traitlets=5.2.1.post0=pypi_0
typer=0.4.1=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py38h06a4308_0
wasabi=0.9.1=pyhd8ed1ab_0
wcwidth=0.2.5=pypi_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py38h0a891b7_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
